src/requirements/gpu_requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.17.0
h5py==2.8.0
idna==2.8
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
matplotlib2tikz==0.6.18
nbformat==4.4.0
numpy==1.15.4
pandas==0.23.4
patsy==0.5.1
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
statsmodels==0.9.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
xlrd==1.2.0
